cypress tests, testt-app and jest test cases moved into single root folder
