+++ tests/jest/mocks/IntersectionObserverMock.js
