+++ tests/test-app/public/favicon.ico
